move Event type mapping to new module
